chore: Rename analysis file from baseline.js -> analyze.js
