+++ lib/reporter/chart.js
@@ -1,6 +1,6 @@
 const { platform, arch, availableParallelism, totalmem } = require("node:os");
 const { styleText } = require("../utils/styleText");
-const { analyze } = require("../utils/baseline.js");
+const { analyze } = require("../utils/analyze.js");
 
 /**
  * Draws a bar chart representation of a benchmark result
