+++ lib/reporter/pretty.js
@@ -1,7 +1,7 @@
 const { styleText } = require("../utils/styleText");
 const { timer } = require("../clock");
 const os = require("node:os");
-const { analyze } = require("../utils/baseline");
+const { analyze } = require("../utils/analyze.js");
 
 const formatter = Intl.NumberFormat(undefined, {
 	notation: "standard",
