+++ lib/reporter/text.js
@@ -1,6 +1,6 @@
 const { styleText } = require("../utils/styleText");
 const { timer } = require("../clock");
-const { analyze } = require("../utils/baseline.js");
+const { analyze } = require("../utils/analyze.js");
 
 const formatter = Intl.NumberFormat(undefined, {
 	notation: "standard",
