+++ lib/utils/analyze.js
